run node via binary in service file - code reworked

+++ test/qnode_service_update_bin.sh
@@ -1,89 +1,197 @@
 #!/bin/bash
 
+#Node version is not used - execution via release_autorun 
+#Comment out for automatic creation of the node version
+#NODE_VERSION=1.4.21.1
 
-# Step 0: Welcome
-echo "✨ Welcome! This script will update your Quilibrium node when running it as a service. ✨"
-echo "It will run the node directly form the binary"
-echo "It will update your service file without changing your customizations"
-echo ""
-echo "Made with 🔥 by LaMat - https://quilibrium.one"
-echo "====================================================================================="
-echo ""
-echo "Processing... ⏳"
-sleep 7  # Add a 7-second delay
+#Comment out for automatic creation of the qclient version
+#QCLIENT_VERSION=1.4.19.1
 
-#===========================
-# Set variables
-#===========================
+cat << "EOF"
 
-# Set service file path
-SERVICE_FILE="/lib/systemd/system/ceremonyclient.service"
-# User working folder
-HOME=$(eval echo ~$USER)
-# Node path
-NODE_PATH="$HOME/ceremonyclient/node"
+                    QQQQQQQQQ       1111111   
+                  QQ:::::::::QQ    1::::::1   
+                QQ:::::::::::::QQ 1:::::::1   
+               Q:::::::QQQ:::::::Q111:::::1   
+               Q::::::O   Q::::::Q   1::::1   
+               Q:::::O     Q:::::Q   1::::1   
+               Q:::::O     Q:::::Q   1::::1   
+               Q:::::O     Q:::::Q   1::::l   
+               Q:::::O     Q:::::Q   1::::l   
+               Q:::::O     Q:::::Q   1::::l   
+               Q:::::O  QQQQ:::::Q   1::::l   
+               Q::::::O Q::::::::Q   1::::l   
+               Q:::::::QQ::::::::Q111::::::111
+                QQ::::::::::::::Q 1::::::::::1
+                  QQ:::::::::::Q  1::::::::::1
+                    QQQQQQQQ::::QQ111111111111
+                            Q:::::Q           
+                             QQQQQQ  QUILIBRIUM.ONE                                                                                                                                
 
-#===========================
-# Check if ceremonyclient directory exists
-#===========================
-HOME=$(eval echo ~$USER)
-CEREMONYCLIENT_DIR="$HOME/ceremonyclient"
+                              
+===========================================================================
+                       ✨ QNODE SERVICE UPDATER ✨
+===========================================================================
+This script will update your Quilibrium node when running it as a service.
+It will run your node from the release_autostart.sh file.
 
-if [ ! -d "$CEREMONYCLIENT_DIR" ]; then
-    echo "❌ Error: You don't have a node installed yet. Nothing to update. Exiting..."
+Follow the guide at https://docs.quilibrium.one
+
+Made with 🔥 by LaMat - https://quilibrium.one
+===========================================================================
+
+Processing... ⏳
+
+EOF
+
+sleep 5  # Add a 7-second delay
+
+#==========================
+# INSTALL APPS
+#==========================
+
+# Function to check and install a package
+check_and_install() {
+    if ! command -v $1 &> /dev/null
+    then
+        echo "$1 could not be found"
+        echo "⏳ Installing $1..."
+        su -c "apt install $1 -y"
+        echo
+    else
+        echo "✅ $1 is installed"
+        echo
+    fi
+}
+
+# For DEBIAN OS - Check if sudo, git, and curl are installed
+check_and_install sudo
+check_and_install git
+check_and_install curl
+
+
+#==========================
+# CREATE PATH VARIABLES
+#==========================
+
+# Determine the ExecStart line based on the architecture
+ARCH=$(uname -m)
+OS=$(uname -s)
+
+# Determine node latest version
+# Check if NODE_VERSION is empty
+if [ -z "$NODE_VERSION" ]; then
+    NODE_VERSION=$(curl -s https://releases.quilibrium.com/release | grep -E "^node-[0-9]+(\.[0-9]+)*" | grep -v "dgst" | sed 's/^node-//' | cut -d '-' -f 1 | head -n 1)
+    if [ -z "$NODE_VERSION" ]; then
+        echo "❌ Error: Unable to determine NODE_VERSION automatically."
+        echo "The script cannot proceed without a correct node version number." 
+        echo "Please try the manual step by step installation instead:"
+        echo "https://docs.quilibrium.one/start/tutorials/node-step-by-step-installation"
+        echo
+        exit 1
+    else
+        echo "✅ Automatically determined NODE_VERSION: $NODE_VERSION"
+    fi
+else
+    echo "✅ Using specified NODE_VERSION: $NODE_VERSION"
+fi
+
+# Determine qclient latest version
+# Check if QCLIENT_VERSION is empty
+if [ -z "$QCLIENT_VERSION" ]; then
+    QCLIENT_VERSION=$(curl -s https://releases.quilibrium.com/qclient-release | grep -E "^qclient-[0-9]+(\.[0-9]+)*" | sed 's/^qclient-//' | cut -d '-' -f 1 |  head -n 1)
+    if [ -z "$QCLIENT_VERSION" ]; then
+        echo "⚠️ Warning: Unable to determine QCLIENT_VERSION automatically. Continuing without it."
+        echo "The script won't be able to install the qclient, but it will still install your node."
+        echo "You can install the qclient later manually if you need to."
+        echo
+        sleep 1
+    else
+        echo "✅ Automatically determined QCLIENT_VERSION: $QCLIENT_VERSION"
+    fi
+else
+    echo "✅ Using specified QCLIENT_VERSION: $QCLIENT_VERSION"
+fi
+
+echo
+
+# Determine the node binary name based on the architecture and OS
+if [ "$ARCH" = "x86_64" ]; then
+    if [ "$OS" = "Linux" ]; then
+        NODE_BINARY="node-$NODE_VERSION-linux-amd64"
+        GO_BINARY="go1.22.4.linux-amd64.tar.gz"
+        [ -n "$QCLIENT_VERSION" ] && QCLIENT_BINARY="qclient-$QCLIENT_VERSION-linux-amd64"
+    elif [ "$OS" = "Darwin" ]; then
+        NODE_BINARY="node-$NODE_VERSION-darwin-amd64"
+        GO_BINARY="go1.22.4.darwin-amd64.tar.gz"
+        [ -n "$QCLIENT_VERSION" ] && QCLIENT_BINARY="qclient-$QCLIENT_VERSION-darwin-amd64"
+    fi
+elif [ "$ARCH" = "aarch64" ]; then
+    if [ "$OS" = "Linux" ]; then
+        NODE_BINARY="node-$NODE_VERSION-linux-arm64"
+        GO_BINARY="go1.22.4.linux-arm64.tar.gz"
+        [ -n "$QCLIENT_VERSION" ] && QCLIENT_BINARY="qclient-$QCLIENT_VERSION-linux-arm64"
+    elif [ "$OS" = "Darwin" ]; then
+        NODE_BINARY="node-$NODE_VERSION-darwin-arm64"
+        GO_BINARY="go1.22.4.darwin-arm64.tar.gz"
+        [ -n "$QCLIENT_VERSION" ] && QCLIENT_BINARY="qclient-$QCLIENT_VERSION-darwin-arm64"
+    fi
+else
+    echo "❌ Error: Unsupported system architecture ($ARCH) or operating system ($OS)."
     exit 1
 fi
 
-#===========================
-# Stop the ceremonyclient service if it exists
-#===========================
-echo "⏳ Stopping the ceremonyclient service if it exists..."
-if systemctl is-active --quiet ceremonyclient; then
-    if sudo systemctl stop ceremonyclient; then
-        echo "🔴 Service stop command issued."
-    else
-        echo "❌ Failed to issue stop command for ceremonyclient service." >&2
-    fi
+#==========================
+# GO UPGRADE
+#==========================
 
-    sleep 1
-
-    # Verify the service has stopped
-    if systemctl is-active --quiet ceremonyclient; then
-        echo "⚠️ Service is still running. Attempting to stop it forcefully..."
-        if sudo systemctl kill ceremonyclient; then
-            sleep 1
-            if systemctl is-active --quiet ceremonyclient; then
-                echo "❌ Service could not be stopped forcefully." >&2
-            else
-                echo "✅ Service stopped forcefully."
-            fi
-        else
-            echo "❌ Failed to force stop the ceremonyclient service." >&2
-        fi
-    else
-        echo "✅ Service stopped successfully."
-    fi
+# Check the currently installed Go version
+if go version &>/dev/null; then
+    INSTALLED_VERSION=$(go version | awk '{print $3}' | sed 's/go//')
 else
-    echo "ℹ️ Ceremonyclient service is not running or does not exist."
+    INSTALLED_VERSION="none"
 fi
 
+# If the installed version is not 1.22.4, proceed with the installation
+if [ "$INSTALLED_VERSION" != "1.22.4" ]; then
+    echo "⏳ Current Go version is $INSTALLED_VERSION. Proceeding with installation of Go 1.22.4..."
+
+    # Download and install Go
+    wget https://go.dev/dl/$GO_BINARY > /dev/null 2>&1 || echo "Failed to download Go!"
+    sudo tar -xvf $GO_BINARY > /dev/null 2>&1 || echo "Failed to extract Go!"
+    sudo rm -rf /usr/local/go || echo "Failed to remove existing Go!"
+    sudo mv go /usr/local || echo "Failed to move Go!"
+    sudo rm $GO_BINARY || echo "Failed to remove downloaded archive!"
+    
+    echo "✅ Go 1.22.4 has been installed successfully."
+    echo
+else
+    echo "✅ Go version 1.22.4 is already installed. No action needed."
+    echo
+fi
+
+#==========================
+# NODE UPDATE
+#==========================
+
+# Stop the ceremonyclient service if it exists
+echo "⏳ Stopping the ceremonyclient service if it exists..."
+if systemctl is-active --quiet ceremonyclient && service ceremonyclient stop; then
+    echo "🔴 Service stopped successfully."
+    echo
+else
+    echo "❌ Ceremonyclient service either does not exist or could not be stopped." >&2
+    echo
+fi
 sleep 1
 
-#===========================
-# Move to the ceremonyclient directory
-#===========================
-echo "Moving to the ceremonyclient directory..."
-cd ~/ceremonyclient || { echo "❌ Error: Directory ~/ceremonyclient does not exist."; exit 1; }
-
-#===========================
 # Discard local changes in release_autorun.sh
-#===========================
-echo "✅ Discarding local changes in release_autorun.sh..."
-git checkout -- node/release_autorun.sh
+# echo "✅ Discarding local changes in release_autorun.sh..."
+# echo
+# git checkout -- node/release_autorun.sh
 
-#===========================
-# Download Binary
-#===========================
+# Set the remote URL and download
+echo "⏳ Downloading new release v$NODE_VERSION"
 cd  ~/ceremonyclient
 git remote set-url origin https://github.com/QuilibriumNetwork/ceremonyclient.git
 #git remote set-url origin https://source.quilibrium.com/quilibrium/ceremonyclient.git || git remote set-url origin https://git.quilibrium-mirror.ch/agostbiro/ceremonyclient.git
@@ -91,44 +199,60 @@ git checkout main
 git branch -D release
 git pull
 git checkout release
+echo "✅ Downloaded the latest changes successfully."
+echo
 
+#==========================
+# QCLIENT UPDATE
+#==========================
 
-sleep 1
+# Building qClient binary
+if [ -n "$QCLIENT_BINARY" ]; then
+    echo "⏳ Downloading qClient..."
+    sleep 1  # Add a 1-second delay
+    cd ~/ceremonyclient/client
 
-# Set the version number
-VERSION=$(cat $NODE_PATH/config/version.go | grep -A 1 "func GetVersion() \[\]byte {" | grep -Eo '0x[0-9a-fA-F]+' | xargs printf "%d.%d.%d")
-
-echo "✅ Downloaded the latest changes successfully - V $VERSION "
-
-#===========================
-# Determine the ExecStart line based on the architecture
-#===========================
-
-# Get the system architecture
-ARCH=$(uname -m)
-
-if [ "$ARCH" = "x86_64" ]; then
-    EXEC_START="$NODE_PATH/node-$VERSION-linux-amd64"
-elif [ "$ARCH" = "aarch64" ]; then
-    EXEC_START="$NODE_PATH/node-$VERSION-linux-arm64"
-elif [ "$ARCH" = "arm64" ]; then
-    EXEC_START="$NODE_PATH/node-$VERSION-darwin-arm64"
+    if ! wget https://releases.quilibrium.com/$QCLIENT_BINARY; then
+        echo "❌ Error: Failed to download qClient binary."
+        echo "Your node will still work, but you'll need to install the qclient manually later if needed."
+        echo
+    else
+        mv $QCLIENT_BINARY qclient
+        chmod +x qclient
+        echo "✅ qClient binary downloaded successfully."
+        echo
+    fi
 else
-    echo "❌ Unsupported architecture: $ARCH"
-    exit 1
+    echo "ℹ️ Skipping qClient download as QCLIENT_BINARY could not be determined earlier."
+    echo "Your node will still work, but you'll need to install the qclient manually later if needed."
+    echo
 fi
 
-sleep 1
+#==========================
+# SERVICE UPDATE
+#==========================
 
-#===========================
-# Re-Create or Update Ceremonyclient Service
-#===========================
-echo "🔧 Checking Ceremonyclient Service..."
-sleep 2  # Add a 2-second delay
+#Set variables
+HOME=$(eval echo ~$HOME_DIR)
 
-if [ ! -f "$SERVICE_FILE" ]; then
-    echo "📝 Service file does not exist. Creating new ceremonyclient service file..."
-    if ! sudo tee "$SERVICE_FILE" > /dev/null <<EOF
+NODE_PATH="$HOME/ceremonyclient/node"
+EXEC_START="$NODE_PATH/$NODE_BINARY"
+
+SERVICE_FILE="/lib/systemd/system/ceremonyclient.service"
+
+# Check if the service file exists and contains the specific ExecStart line with any parameters
+if [ -f "$SERVICE_FILE" ] && grep -q "ExecStart=/root/scripts/qnode_cluster_run.sh" "$SERVICE_FILE"; then
+    echo "⚠️ You are running a cluster. Skipping service file update..."
+    echo
+else
+    echo "⏳ Rebuilding Ceremonyclient Service..."
+    echo
+    sleep 1
+
+    if [ ! -f "$SERVICE_FILE" ]; then
+        echo "⏳ Creating new ceremonyclient service file..."
+        echo
+        if ! sudo tee "$SERVICE_FILE" > /dev/null <<EOF
 [Unit]
 Description=Ceremony Client Go App Service
 
@@ -138,68 +262,111 @@ Restart=always
 RestartSec=5s
 WorkingDirectory=$NODE_PATH
 ExecStart=$EXEC_START
+KillSignal=SIGINT
+TimeoutStopSec=30s
 
 [Install]
 WantedBy=multi-user.target
 EOF
-    then
-        echo "❌ Error: Failed to create ceremonyclient service file." >&2
-        exit 1
-    fi
-else
-    echo "🔍 Checking existing ceremonyclient service file..."
-
-    # Replace the existing lines with new values
-    if sudo grep -q "WorkingDirectory=" "$SERVICE_FILE"; then
-        sudo sed -i "s|^WorkingDirectory=.*|WorkingDirectory=$NODE_PATH|" "$SERVICE_FILE"
+        then
+            echo "❌ Error: Failed to create ceremonyclient service file." >&2
+            exit 1
+        fi
     else
-        echo "WorkingDirectory not found in the service file. Adding it."
-        sudo sed -i "/\[Service\]/a WorkingDirectory=$NODE_PATH" "$SERVICE_FILE"
+        echo "⏳ Checking existing ceremonyclient service file..."
+        echo
+        
+        # Check if the required lines exist
+        if ! grep -q "WorkingDirectory=$NODE_PATH" "$SERVICE_FILE" || \
+           ! grep -q "ExecStart=$EXEC_START" "$SERVICE_FILE" || \
+           ! grep -q "KillSignal=SIGINT" "$SERVICE_FILE" || \
+           ! grep -q "TimeoutStopSec=30s" "$SERVICE_FILE"; then
+            echo "⏳ Updating existing ceremonyclient service file..."
+            echo
+            # Replace or add the lines with new values
+            sudo sed -i "s|WorkingDirectory=.*|WorkingDirectory=$NODE_PATH|" "$SERVICE_FILE"
+            sudo sed -i "s|ExecStart=.*|ExecStart=$EXEC_START|" "$SERVICE_FILE"
+            
+            # Add KillSignal and TimeoutStopSec if they don't exist
+            if ! grep -q "KillSignal=SIGINT" "$SERVICE_FILE"; then
+                sudo sed -i '/\[Service\]/a KillSignal=SIGINT' "$SERVICE_FILE"
+            fi
+            if ! grep -q "TimeoutStopSec=30s" "$SERVICE_FILE"; then
+                sudo sed -i '/\[Service\]/a TimeoutStopSec=30s' "$SERVICE_FILE"
+            fi
+        else
+            echo "✅ No changes needed."
+        fi
     fi
-
-    if sudo grep -q "ExecStart=" "$SERVICE_FILE"; then
-        sudo sed -i "s|^ExecStart=.*|ExecStart=$EXEC_START|" "$SERVICE_FILE"
-    else
-        echo "ExecStart not found in the service file. Adding it."
-        sudo sed -i "/\[Service\]/a ExecStart=$EXEC_START" "$SERVICE_FILE"
-    fi
-
-    echo "✅ Ceremonyclient service file updated."
-fi  
-
-sleep 1  # Add a 1-second delay
-
-#===========================
-# Remove the SELF_TEST file
-#===========================
-if [ -f "$NODE_PATH/.config/SELF_TEST" ]; then
-    echo "🗑️ Removing SELF_TEST file..."
-    if rm "$NODE_PATH/.config/SELF_TEST"; then
-        echo "✅ SELF_TEST file removed successfully."
-    else
-        echo "❌ Error: Failed to remove SELF_TEST file." >&2
-        exit 1
-    fi
-else
-    echo "ℹ️ No SELF_TEST file found at $NODE_PATH/.config/SELF_TEST."
 fi
-sleep 1  # Add a 1-second delay
 
-#===========================
-# Start the ceremonyclient service
-#===========================
+# The script continues here with any code that should run after the conditional block
+echo "Continuing with the rest of the script..."
+
+echo
+
+#==========================
+# CONFIG FILE UPDATE for "REWARDS TO GOOGLE SHEET SCRIPT"
+#==========================
+
+# Function to update config file
+update_config_file() {
+    local config_file=$1
+    echo "✅ Checking node version in config file '$(basename "$config_file")'."
+    sleep 1
+    # Get the current NODE_BINARY from the config file
+    config_node_binary=$(grep "NODE_BINARY=" "$config_file" | cut -d '=' -f 2)
+    # Compare NODE_BINARY values
+    if [ "$config_node_binary" = "$NODE_BINARY" ]; then
+        echo "NODE_BINARY values match. No update needed."
+    else
+        echo "⏳ NODE_BINARY values differ. Updating config file..."
+        
+        # Update the config file
+        sed -i "s|NODE_BINARY=.*|NODE_BINARY=$NODE_BINARY|" "$config_file"
+        
+        if [ $? -eq 0 ]; then
+            echo "✅ Config file updated successfully."
+        else
+            echo "❌ Failed to update config file. Continuing to next step..."
+        fi
+    fi
+}
+
+# Array of config files
+config_files=(
+    "~/scripts/qnode_rewards_to_gsheet.config"
+    "~/scripts/qnode_rewards_to_gsheet_2.config"
+)
+
+# Loop through config files
+for config_file in "${config_files[@]}"; do
+    if [ -f "$config_file" ]; then
+        update_config_file "$config_file"
+    else
+        echo "Config file not found: $config_file"
+        echo "This is not a problem (it's not related to your node). Continuing to next file..."
+    fi
+done
+
+echo "All config files processed."
+
+echo
+
+#==========================
+# START NODE VIA SERVICE
+#==========================
+
 echo "✅ Starting Ceremonyclient Service"
-sleep 2  # Add a 2-second delay
-systemctl daemon-reload
-systemctl enable ceremonyclient
-service ceremonyclient start
+sudo systemctl daemon-reload
+sudo systemctl enable ceremonyclient
+sudo service ceremonyclient start
 
-#===========================
-# Showing the node logs
-#===========================
-echo ""
-echo "🌟Your Qnode is now updated to V $VERSION!"
-echo ""
-echo "⏳ Showing the node log... (Hit Ctrl+C to exit log)"
-sleep 1
+# Showing the node version and logs
+echo "🌟Your node is now updated to v$NODE_VERSION !"
+echo
+echo "⏳ Showing the node log... (CTRL+C to exit)"
+echo
+echo
+sleep 2
 sudo journalctl -u ceremonyclient.service -f --no-hostname -o cat
